Move tensor into init.lua
